+++ tests/test_config.py
@@ -154,6 +154,27 @@ def test_config_merging_missing():
     assert result["ignore_names"] == ["name1"]
 
 
+def test_config_merging_toml_paths_only():
+    """
+    If we have paths in the TOML but not on the CLI, the TOML paths should be
+    used.
+    """
+    toml = StringIO(
+        dedent(
+            """\
+        [tool.vulture]
+        paths = ["path1", "path2"]
+        """
+        )
+    )
+    cliargs = [
+        "--exclude=test_*.py",
+    ]
+    result = make_config(cliargs, toml)
+    assert result["paths"] == ["path1", "path2"]
+    assert result["exclude"] == ["test_*.py"]
+
+
 def test_invalid_config_options_output():
     """
     If the config file contains unknown options we want to abort.