+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   changed git signing key to correct verification
 
 * Only parse format strings when being used with `locals()` (jingw, #225).
+* Don't override paths in pyproject.toml with empty CLI paths (bcbnz, #228).
+* Run continuous integration tests for Python 3.9 (ju-sh, #232).
+* Use pathlib internally (ju-sh, #226).
 
 # 2.1 (2020-08-19)